refactor: split production and military units panels into scoped parts (#3878)
Continue Phase 3 panel decomposition by extracting available/allocation
subpanels and military detail-row widgets into part files, keeping main
library files under 300 lines without changing panel behavior.
