+++ app/lib/features/game/widgets/production_panel.dart
@@ -22,6 +22,9 @@ import 'production_available_grid.dart';
 import 'production_labour_helpers.dart';
 import 'production_labour_section.dart';
 
+part 'production_panel_support_available.dart';
+part 'production_panel_support_allocation.dart';
+
 /// Public layout key planted on the narrow (`< [kNarrowBreakpoint]` dp) variant
 /// of [ProductionPanel] so widget tests and Widgetbook pinning can confirm
 /// that the screen has selected its `_ProductionPanelNarrowLayout` branch
@@ -222,447 +225,3 @@ class _ProductionPanelWideLayout extends StatelessWidget {
     );
   }
 }
-
-class _AvailableSubpanel extends StatelessWidget {
-  const _AvailableSubpanel({
-    required this.game,
-    required this.player,
-    required this.effectiveLabour,
-    required this.inputCommodityIds,
-    required this.outputCommodityIds,
-    required this.netDeltasByCommodity,
-    required this.l10n,
-    this.onOpenCommodityBreakdown,
-    this.currentOrders,
-    this.labourCallbacks,
-    this.canEditLabour = false,
-  });
-
-  final Game game;
-  final Player player;
-  final int effectiveLabour;
-  final Set<String> inputCommodityIds;
-  final Set<String> outputCommodityIds;
-  final Map<String, int> netDeltasByCommodity;
-  final AppLocalizations l10n;
-  final VoidCallback? onOpenCommodityBreakdown;
-  final Orders? currentOrders;
-  final ProductionLabourCallbacks? labourCallbacks;
-  final bool canEditLabour;
-
-  /// Quantity shown in Available commodity cells for tradeable stock.
-  ///
-  /// Matches the Trade Market tab `(N)` sellable readout:
-  /// `sellableHeadroomByCommodityId` (offer cap minus staged offers).
-  /// Riches and zero-stockpile commodities fall back to raw stockpile.
-  int _displayedStockpileQuantity(
-    Commodity c,
-    Map<CommodityId, int> sellableByCommodityId,
-  ) {
-    final headroom = sellableByCommodityId[c.id];
-    if (headroom != null) return headroom;
-    return player.stockpile.quantityOf(c.id);
-  }
-
-  Widget _buildCommodityCell(Commodity c, int qty, int change) {
-    final name = c.displayName ?? c.id;
-    return CtResourceCell(
-      key: ValueKey<String>('production_available_cell_${c.id}'),
-      iconBuilder: (_) => ResourceIcon(
-        commodityId: c.id,
-        size: CtResourceCell.leadingIconSize,
-      ),
-      name: name,
-      quantity: qty,
-      delta: change == 0 ? null : change,
-    );
-  }
-
-  Widget _buildCommodityGrid(
-    List<Commodity> commodities,
-    Map<String, int> netChanges,
-    Map<CommodityId, int> sellableByCommodityId,
-  ) {
-    return AvailableCellGrid(
-      key: ValueKey<String>(
-        'production_available_commodity_grid_${commodities.map((c) => c.id).join('_')}',
-      ),
-      columnCount: kProductionAvailableCommodityGridColumns,
-      cells: <Widget>[
-        for (final commodity in commodities)
-          _buildCommodityCell(
-            commodity,
-            _displayedStockpileQuantity(commodity, sellableByCommodityId),
-            netChanges[commodity.id] ?? 0,
-          ),
-      ],
-    );
-  }
-
-  Widget _buildWorkerCell(String workerType, int count) {
-    return CtResourceCell(
-      key: ValueKey<String>('production_available_worker_$workerType'),
-      iconBuilder: (_) => WorkerIcon(
-        workerType: workerType,
-        size: CtResourceCell.leadingIconSize,
-      ),
-      name: _workerDisplayName(workerType),
-      quantity: count,
-    );
-  }
-
-  String _workerDisplayName(String workerType) {
-    switch (workerType) {
-      case 'peasant':
-        return l10n.production_workers_peasants;
-      case 'apprentice':
-        return l10n.production_workers_apprentices;
-      case 'journeyman':
-        return l10n.production_workers_journeymen;
-      case 'master':
-        return l10n.production_workers_masters;
-      default:
-        return workerType;
-    }
-  }
-
-  List<Widget> _buildFoodSection(
-    List<Commodity> availableFood,
-    Map<String, int> netChanges,
-    Map<CommodityId, int> sellableByCommodityId,
-  ) {
-    if (availableFood.isEmpty) {
-      return const <Widget>[];
-    }
-    return <Widget>[
-      CtSectionLabel(l10n.production_food),
-      const SizedBox(height: 6),
-      _buildCommodityGrid(availableFood, netChanges, sellableByCommodityId),
-      CtGap.ml,
-    ];
-  }
-
-  List<Widget> _buildMaterialsSection(
-    List<Commodity> rawMaterials,
-    List<Commodity> manufactured,
-    Map<String, int> netChanges,
-    Map<CommodityId, int> sellableByCommodityId,
-  ) {
-    final children = <Widget>[
-      CtSectionLabel(l10n.production_rawMaterials),
-      const SizedBox(height: 6),
-      _buildCommodityGrid(rawMaterials, netChanges, sellableByCommodityId),
-    ];
-    if (manufactured.isNotEmpty) {
-      children.addAll([
-        CtGap.ml,
-        CtSectionLabel(l10n.production_manufactured),
-        const SizedBox(height: 6),
-        _buildCommodityGrid(manufactured, netChanges, sellableByCommodityId),
-      ]);
-    }
-    return children;
-  }
-
-  List<Widget> _buildWorkerSection(ThemeData theme) {
-    final children = <Widget>[
-      CtGap.ml,
-      CtSectionLabel(l10n.production_workers),
-      const SizedBox(height: 6),
-      AvailableCellGrid(
-        key: kProductionAvailableWorkerGridKey,
-        columnCount: kProductionAvailableWorkerGridColumns,
-        cells: <Widget>[
-          _buildWorkerCell('peasant', player.workerPool.peasants),
-          _buildWorkerCell('apprentice', player.workerPool.apprentices),
-          _buildWorkerCell('journeyman', player.workerPool.journeymen),
-          _buildWorkerCell('master', player.workerPool.masters),
-        ],
-      ),
-      CtGap.m,
-      EffectiveLabourTotal(
-        text: l10n.production_effectiveLabour(effectiveLabour),
-        theme: theme,
-      ),
-    ];
-    if (currentOrders != null && labourCallbacks != null) {
-      children.addAll(<Widget>[
-        CtGap.ml,
-        CtSectionLabel(l10n.production_labourControlsSectionLabel),
-        const SizedBox(height: 6),
-        ProductionLabourSection(
-          player: player,
-          currentOrders: currentOrders!,
-          canEdit: canEditLabour,
-          callbacks: labourCallbacks!,
-        ),
-      ]);
-    }
-    return children;
-  }
-
-  Widget _buildHeader(ThemeData theme) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        Expanded(
-          child: Text(
-            l10n.production_available,
-            style: theme.textTheme.titleSmall,
-          ),
-        ),
-        if (onOpenCommodityBreakdown != null)
-          CtActionTextButton(
-            onPressed: onOpenCommodityBreakdown,
-            label: l10n.production_breakdown,
-          ),
-      ],
-    );
-  }
-
-  List<Widget> _buildBodyChildren(ThemeData theme) {
-    final netChanges = netDeltasByCommodity;
-    final sellableByCommodityId = sellableHeadroomByCommodityId(
-      game: game,
-      playerId: player.id,
-      orders: currentOrders ?? const Orders(),
-    );
-    final rawMaterials = CommodityCatalog.all
-        .where(
-          (c) =>
-              c.category == CommodityCategory.rawMaterial &&
-              inputCommodityIds.contains(c.id),
-        )
-        .toList();
-    final manufactured = CommodityCatalog.all
-        .where((c) => c.category == CommodityCategory.manufactured)
-        .toList();
-    final availableFood = CommodityCatalog.all
-        .where((c) => c.category == CommodityCategory.food)
-        .toList();
-    return <Widget>[
-      _buildHeader(theme),
-      CtGap.m,
-      ..._buildFoodSection(
-        availableFood,
-        netChanges,
-        sellableByCommodityId,
-      ),
-      ..._buildMaterialsSection(
-        rawMaterials,
-        manufactured,
-        netChanges,
-        sellableByCommodityId,
-      ),
-      ..._buildWorkerSection(theme),
-    ];
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return CtPanel(
-      padding: const EdgeInsets.all(CtSpacing.ml),
-      child: SingleChildScrollView(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: _buildBodyChildren(theme),
-        ),
-      ),
-    );
-  }
-}
-
-class _AllocationSubpanel extends StatelessWidget {
-  const _AllocationSubpanel({
-    required this.player,
-    required this.effectiveLabour,
-    required this.desiredOutputByRecipe,
-    required this.onDesiredOutputChanged,
-    required this.l10n,
-  });
-
-  final Player player;
-  final int effectiveLabour;
-  final Map<String, int> desiredOutputByRecipe;
-  final ValueChanged<Map<String, int>> onDesiredOutputChanged;
-  final AppLocalizations l10n;
-
-  Widget _buildRecipeLabel(
-    ProductionRecipe recipe,
-    ThemeData theme,
-    bool locked,
-  ) {
-    final outputCommodity = CommodityCatalog.byId[recipe.outputCommodityId];
-    final outputName = outputCommodity?.displayName ?? recipe.outputCommodityId;
-    final inputParts = recipe.inputQuantities.entries
-        .map((e) {
-          final comm = CommodityCatalog.byId[e.key];
-          final name = comm?.displayName ?? e.key;
-          return '$name ×${e.value}';
-        })
-        .join(', ');
-    final label = '$outputName ($inputParts)';
-
-    final hasOutputIcon =
-        outputCommodity != null && ResourceIcon.hasIcon(outputCommodity.id);
-
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        if (hasOutputIcon) ...[
-          ResourceIcon(commodityId: recipe.outputCommodityId, size: 14),
-          const SizedBox(width: 4),
-        ],
-        Flexible(
-          child: Text(
-            label,
-            style: theme.textTheme.labelMedium,
-            overflow: TextOverflow.ellipsis,
-          ),
-        ),
-        if (locked) ...[
-          const SizedBox(width: 4),
-          Text(
-            l10n.production_recipeLocked,
-            style: theme.textTheme.labelMedium?.copyWith(
-              color: EditorialMonoclePalette.muted,
-            ),
-          ),
-        ],
-      ],
-    );
-  }
-
-  int _computeTotalRequiredLabour() {
-    return desiredOutputByRecipe.entries.fold<int>(0, (sum, entry) {
-      final recipe = ProductionRecipesCatalog.byId[entry.key];
-      if (recipe == null) {
-        return sum;
-      }
-      return sum + entry.value * recipe.labourPerOutput;
-    });
-  }
-
-  Widget _buildHeader(ThemeData theme) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        Expanded(
-          child: Text(
-            l10n.production_allocation,
-            style: theme.textTheme.titleSmall,
-          ),
-        ),
-        CtDangerTextButton(
-          key: const ValueKey<String>('production_allocation_reset_button'),
-          onPressed: () => onDesiredOutputChanged({}),
-          label: l10n.common_reset,
-          tooltip: l10n.common_reset,
-        ),
-      ],
-    );
-  }
-
-  /// Builds the recipe rows interleaved with `CtBrassDivider`s per
-  /// `SPEC/ui/production-panel.md` § Allocation row chrome — exactly
-  /// `N - 1` dividers between consecutive rows, none at the leading or
-  /// trailing edges (`Refs #2862` S3 / R13).
-  List<Widget> _buildAllocationRows(ThemeData theme) {
-    final recipes = ProductionRecipesCatalog.all;
-    if (recipes.isEmpty) {
-      return const <Widget>[];
-    }
-    final widgets = <Widget>[];
-    for (int i = 0; i < recipes.length; i++) {
-      if (i > 0) {
-        widgets.add(
-          const Padding(
-            padding: EdgeInsets.symmetric(vertical: 4),
-            child: CtBrassDivider(),
-          ),
-        );
-      }
-      final recipe = recipes[i];
-      final locked = !ProductionRecipesCatalog.isRecipeAvailableForPlayer(
-        recipe,
-        player.techUnlocked,
-      );
-      widgets.add(
-        ProductionAllocationRowChrome(
-          key: ValueKey<String>('production_alloc_row_chrome_${recipe.id}'),
-          child: ProductionAllocationRow(
-            key: ValueKey<String>('production_alloc_row_${recipe.id}'),
-            recipe: recipe,
-            player: player,
-            effectiveLabour: effectiveLabour,
-            desiredOutputByRecipe: desiredOutputByRecipe,
-            onDesiredOutputChanged: onDesiredOutputChanged,
-            buildRecipeLabel: (value, isLocked) =>
-                _buildRecipeLabel(value, theme, isLocked),
-            l10n: l10n,
-            theme: theme,
-            locked: locked,
-          ),
-        ),
-      );
-    }
-    return widgets;
-  }
-
-  List<Widget> _buildLabourSummary(
-    ThemeData theme,
-    int totalRequiredLabour,
-    bool labourInsufficient,
-  ) {
-    return <Widget>[
-      CtGap.m,
-      Text(
-        l10n.production_totalLabour(totalRequiredLabour, effectiveLabour),
-        style: theme.textTheme.bodySmall?.copyWith(
-          fontWeight: FontWeight.w600,
-          color: labourInsufficient ? theme.colorScheme.error : null,
-        ),
-      ),
-      if (labourInsufficient)
-        Padding(
-          padding: const EdgeInsets.only(top: 4),
-          child: Text(
-            l10n.production_labourInsufficient,
-            style: theme.textTheme.bodySmall?.copyWith(
-              color: theme.colorScheme.error,
-            ),
-          ),
-        ),
-    ];
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final totalRequiredLabour = _computeTotalRequiredLabour();
-    final labourInsufficient = totalRequiredLabour > effectiveLabour;
-
-    return CtPanel(
-      padding: const EdgeInsets.fromLTRB(12, 12, 20, 12),
-      child: SingleChildScrollView(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            _buildHeader(theme),
-            CtGap.m,
-            ..._buildAllocationRows(theme),
-            ..._buildLabourSummary(
-              theme,
-              totalRequiredLabour,
-              labourInsufficient,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
