+++ app/lib/features/game/widgets/production_panel_support_available.dart
@@ -1,0 +1,254 @@
+/// Available subpanel for production screen. SPEC/ui/production-panel.md.
+
+part of 'production_panel.dart';
+
+class _AvailableSubpanel extends StatelessWidget {
+  const _AvailableSubpanel({
+    required this.game,
+    required this.player,
+    required this.effectiveLabour,
+    required this.inputCommodityIds,
+    required this.outputCommodityIds,
+    required this.netDeltasByCommodity,
+    required this.l10n,
+    this.onOpenCommodityBreakdown,
+    this.currentOrders,
+    this.labourCallbacks,
+    this.canEditLabour = false,
+  });
+
+  final Game game;
+  final Player player;
+  final int effectiveLabour;
+  final Set<String> inputCommodityIds;
+  final Set<String> outputCommodityIds;
+  final Map<String, int> netDeltasByCommodity;
+  final AppLocalizations l10n;
+  final VoidCallback? onOpenCommodityBreakdown;
+  final Orders? currentOrders;
+  final ProductionLabourCallbacks? labourCallbacks;
+  final bool canEditLabour;
+
+  /// Quantity shown in Available commodity cells for tradeable stock.
+  ///
+  /// Matches the Trade Market tab `(N)` sellable readout:
+  /// `sellableHeadroomByCommodityId` (offer cap minus staged offers).
+  /// Riches and zero-stockpile commodities fall back to raw stockpile.
+  int _displayedStockpileQuantity(
+    Commodity c,
+    Map<CommodityId, int> sellableByCommodityId,
+  ) {
+    final headroom = sellableByCommodityId[c.id];
+    if (headroom != null) return headroom;
+    return player.stockpile.quantityOf(c.id);
+  }
+
+  Widget _buildCommodityCell(Commodity c, int qty, int change) {
+    final name = c.displayName ?? c.id;
+    return CtResourceCell(
+      key: ValueKey<String>('production_available_cell_${c.id}'),
+      iconBuilder: (_) => ResourceIcon(
+        commodityId: c.id,
+        size: CtResourceCell.leadingIconSize,
+      ),
+      name: name,
+      quantity: qty,
+      delta: change == 0 ? null : change,
+    );
+  }
+
+  Widget _buildCommodityGrid(
+    List<Commodity> commodities,
+    Map<String, int> netChanges,
+    Map<CommodityId, int> sellableByCommodityId,
+  ) {
+    return AvailableCellGrid(
+      key: ValueKey<String>(
+        'production_available_commodity_grid_${commodities.map((c) => c.id).join('_')}',
+      ),
+      columnCount: kProductionAvailableCommodityGridColumns,
+      cells: <Widget>[
+        for (final commodity in commodities)
+          _buildCommodityCell(
+            commodity,
+            _displayedStockpileQuantity(commodity, sellableByCommodityId),
+            netChanges[commodity.id] ?? 0,
+          ),
+      ],
+    );
+  }
+
+  Widget _buildWorkerCell(String workerType, int count) {
+    return CtResourceCell(
+      key: ValueKey<String>('production_available_worker_$workerType'),
+      iconBuilder: (_) => WorkerIcon(
+        workerType: workerType,
+        size: CtResourceCell.leadingIconSize,
+      ),
+      name: _workerDisplayName(workerType),
+      quantity: count,
+    );
+  }
+
+  String _workerDisplayName(String workerType) {
+    switch (workerType) {
+      case 'peasant':
+        return l10n.production_workers_peasants;
+      case 'apprentice':
+        return l10n.production_workers_apprentices;
+      case 'journeyman':
+        return l10n.production_workers_journeymen;
+      case 'master':
+        return l10n.production_workers_masters;
+      default:
+        return workerType;
+    }
+  }
+
+  List<Widget> _buildFoodSection(
+    List<Commodity> availableFood,
+    Map<String, int> netChanges,
+    Map<CommodityId, int> sellableByCommodityId,
+  ) {
+    if (availableFood.isEmpty) {
+      return const <Widget>[];
+    }
+    return <Widget>[
+      CtSectionLabel(l10n.production_food),
+      const SizedBox(height: 6),
+      _buildCommodityGrid(availableFood, netChanges, sellableByCommodityId),
+      CtGap.ml,
+    ];
+  }
+
+  List<Widget> _buildMaterialsSection(
+    List<Commodity> rawMaterials,
+    List<Commodity> manufactured,
+    Map<String, int> netChanges,
+    Map<CommodityId, int> sellableByCommodityId,
+  ) {
+    final children = <Widget>[
+      CtSectionLabel(l10n.production_rawMaterials),
+      const SizedBox(height: 6),
+      _buildCommodityGrid(rawMaterials, netChanges, sellableByCommodityId),
+    ];
+    if (manufactured.isNotEmpty) {
+      children.addAll([
+        CtGap.ml,
+        CtSectionLabel(l10n.production_manufactured),
+        const SizedBox(height: 6),
+        _buildCommodityGrid(manufactured, netChanges, sellableByCommodityId),
+      ]);
+    }
+    return children;
+  }
+
+  List<Widget> _buildWorkerSection(ThemeData theme) {
+    final children = <Widget>[
+      CtGap.ml,
+      CtSectionLabel(l10n.production_workers),
+      const SizedBox(height: 6),
+      AvailableCellGrid(
+        key: kProductionAvailableWorkerGridKey,
+        columnCount: kProductionAvailableWorkerGridColumns,
+        cells: <Widget>[
+          _buildWorkerCell('peasant', player.workerPool.peasants),
+          _buildWorkerCell('apprentice', player.workerPool.apprentices),
+          _buildWorkerCell('journeyman', player.workerPool.journeymen),
+          _buildWorkerCell('master', player.workerPool.masters),
+        ],
+      ),
+      CtGap.m,
+      EffectiveLabourTotal(
+        text: l10n.production_effectiveLabour(effectiveLabour),
+        theme: theme,
+      ),
+    ];
+    if (currentOrders != null && labourCallbacks != null) {
+      children.addAll(<Widget>[
+        CtGap.ml,
+        CtSectionLabel(l10n.production_labourControlsSectionLabel),
+        const SizedBox(height: 6),
+        ProductionLabourSection(
+          player: player,
+          currentOrders: currentOrders!,
+          canEdit: canEditLabour,
+          callbacks: labourCallbacks!,
+        ),
+      ]);
+    }
+    return children;
+  }
+
+  Widget _buildHeader(ThemeData theme) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        Expanded(
+          child: Text(
+            l10n.production_available,
+            style: theme.textTheme.titleSmall,
+          ),
+        ),
+        if (onOpenCommodityBreakdown != null)
+          CtActionTextButton(
+            onPressed: onOpenCommodityBreakdown,
+            label: l10n.production_breakdown,
+          ),
+      ],
+    );
+  }
+
+  List<Widget> _buildBodyChildren(ThemeData theme) {
+    final netChanges = netDeltasByCommodity;
+    final sellableByCommodityId = sellableHeadroomByCommodityId(
+      game: game,
+      playerId: player.id,
+      orders: currentOrders ?? const Orders(),
+    );
+    final rawMaterials = CommodityCatalog.all
+        .where(
+          (c) =>
+              c.category == CommodityCategory.rawMaterial &&
+              inputCommodityIds.contains(c.id),
+        )
+        .toList();
+    final manufactured = CommodityCatalog.all
+        .where((c) => c.category == CommodityCategory.manufactured)
+        .toList();
+    final availableFood = CommodityCatalog.all
+        .where((c) => c.category == CommodityCategory.food)
+        .toList();
+    return <Widget>[
+      _buildHeader(theme),
+      CtGap.m,
+      ..._buildFoodSection(
+        availableFood,
+        netChanges,
+        sellableByCommodityId,
+      ),
+      ..._buildMaterialsSection(
+        rawMaterials,
+        manufactured,
+        netChanges,
+        sellableByCommodityId,
+      ),
+      ..._buildWorkerSection(theme),
+    ];
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return CtPanel(
+      padding: const EdgeInsets.all(CtSpacing.ml),
+      child: SingleChildScrollView(
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: _buildBodyChildren(theme),
+        ),
+      ),
+    );
+  }
+}
